src/backend/requirements.txt
alembic==1.4.0
bcrypt==3.1.7
cffi==1.14.0
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==3.3.0
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
